auth-service/app/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "user auth service"
authors = ["Denis Shvetsov <den.shvetsov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
redis = "4.4.2"
fastapi = {extras = ["all"], version = "^0.110.0"}
orjson = "^3.9.10"
uvicorn = "^0.25.0"
uvloop = "^0.19.0"
aiohttp = "^3.9.1"
flake8 = "^6.1.0"
black = "^23.12.1"
isort = "^5.13.2"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
werkzeug = "^3.0.1"
psycopg2 = "^2.9.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
pycryptodome = "^3.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
